Be

See merge request s09-webmobile1-sub2/S09P12E201!34

Former-commit-id: c6f5db4eb275e9c68ffb2310883114a735cfb8d1

+++ backend/src/main/java/com/ssafy/api/service/BadgeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.api.model.Badge;
 
 @Repository
-public interface BadgeRepository extends JpaRepository<Badge, Integer>{
+public interface BadgeRepository extends JpaRepository<Badge, Long>{
 
 }
